fix: 페이지 라우터 방식에서 앱 라우터 방식으로 폴더 구조 변경
