+++ package.json
@@ -22,7 +22,7 @@
     "eslint": "8.49.0",
     "eslint-config-next": "13.4.19",
     "next": "13.4.19",
-    "next-auth": "^4.24.11",
+    "next-auth": "^4.24.13",
     "postcss": "8.4.29",
     "react": "18.2.0",
     "react-daum-postcode": "^3.2.0",
