+++ src/app/stores/[id]/edit/page.tsx
@@ -1,35 +1,79 @@
+"use client";
+
 import { CATEGORY_ARR, FOOD_CERTIFY_ARR, STORE_TYPE_ARR } from "@/data/store";
 import axios from "axios";
 import { useForm } from "react-hook-form";
 import { toast } from "react-toastify";
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
 import AddressSearch from "@/components/AddressSearch";
 import { StoreType } from "@/interface";
+import { useQuery } from "react-query";
+import Loader from "@/components/Loader";
 
-export default function StoreNewPage() {
+export default function StoreEditPage({ params }: { params: { id: string } }) {
     const router = useRouter();
+    const id = params.id;
+
+    const fetchStore = async () => {
+        const { data } = await axios(`/api/stores?id=${id}`);
+        return data as StoreType;
+    };
+
     const {
         register,
         handleSubmit,
         setValue,
         formState: { errors },
     } = useForm<StoreType>();
+
+    const {
+        data: store,
+        isFetching,
+        isSuccess,
+        isError,
+    } = useQuery(`store-${id}`, fetchStore, {
+        onSuccess: (data) => {
+            setValue("id", data.id);
+            setValue("name", data.name);
+            setValue("category", data.category);
+            setValue("phone", data.phone);
+            setValue("address", data.address);
+            setValue("lat", data.lat);
+            setValue("lng", data.lng);
+            setValue("foodCertifyName", data.foodCertifyName);
+            setValue("storeType", data.storeType);
+        },
+        refetchOnWindowFocus: false,
+    });
+
+    if (isError) {
+        return (
+            <div className="w-full h-screen mx-auto pt-[10%] text-gray-500 text-center font-semibold">
+                다시 시도해주세요.
+            </div>
+        );
+    }
+
+    if (isFetching) {
+        return <Loader className="mt-[40%]" />;
+    }
+
     return (
         <form
             className="px-4 md:max-w-4xl mx-auto py-8"
             onSubmit={handleSubmit(async (data) => {
                 try {
-                    const result = await axios.post("/api/stores", data);
+                    const result = await axios.put("/api/stores", data);
 
                     if (result.status === 200) {
-                        toast.success("맛집을 등록했습니다.");
+                        toast.success("맛집을 수정했습니다.");
                         router.replace(`/stores/${result.data.id}`);
                     } else {
                         toast.error("다시 시도해주세요.");
                     }
                 } catch (e) {
                     toast.error(
-                        "데이터 생성 중 문제가 생겼습니다. 다시 시도해주세요."
+                        "데이터 수정 중 문제가 생겼습니다. 다시 시도해주세요."
                     );
                 }
             })}
@@ -37,10 +81,10 @@ export default function StoreNewPage() {
             <div className="space-y-12">
                 <div className="border-b border-gray-900/10 pb-12">
                     <h2 className="text-base/7 font-semibold text-gray-900">
-                        맛집 등록
+                        맛집 수정
                     </h2>
                     <p className="mt-1 text-sm/6 text-gray-600">
-                        아래 내용을 입력해서 맛집을 등록해주세요.
+                        아래 내용을 입력해서 맛집을 수정해주세요.
                     </p>
 
                     <div className="mt-10 grid grid-cols-1 gap-x-6 gap-y-8 sm:grid-cols-6">
@@ -194,7 +238,7 @@ export default function StoreNewPage() {
                     type="submit"
                     className="rounded-md bg-indigo-600 px-3 py-2 text-sm font-semibold text-white shadow-xs hover:bg-indigo-500 focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600"
                 >
-                    등록하기
+                    수정하기
                 </button>
             </div>
         </form>
