+++ src/app/stores/[id]/page.tsx
@@ -1,4 +1,6 @@
-import { useRouter } from "next/router";
+"use client";
+
+import { useRouter } from "next/navigation";
 import { useQuery } from "react-query";
 import axios from "axios";
 import { StoreType } from "@/interface";
@@ -11,9 +13,13 @@ import { toast } from "react-toastify";
 import Like from "@/components/Like";
 import Comments from "@/components/comments";
 
-export default function StoreDetailPage() {
+export default function StoreDetailPage({
+    params,
+}: {
+    params: { id: string };
+}) {
     const router = useRouter();
-    const { id } = router.query;
+    const id = params.id;
     const { status } = useSession();
 
     const FetchStore = async () => {
