+++ src/app/users/mypage/page.tsx
@@ -1,15 +1,15 @@
+"use client";
+
 import CommentList from "@/components/comments/CommentList";
 import Pagination from "@/components/Pagination";
 import { CommentApiResponse } from "@/interface";
 import axios from "axios";
 import { signOut, useSession } from "next-auth/react";
-import { useRouter } from "next/router";
 import { useQuery } from "react-query";
 
-export default function MyPage() {
+export default function MyPage({ params }: { params: { page: string } }) {
     const { data: session } = useSession();
-    const router = useRouter();
-    const { page = "1" }: any = router.query;
+    const page = params.page || "1";
 
     const fetchComments = async () => {
         const { data } = await axios.get(
