+++ src/components/Map.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 /*global kakao*/
 import Script from "next/script";
 import { useRecoilValue, useSetRecoilState } from "recoil";
