+++ src/components/Marker.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import { mapState } from "@/atom";
 import { StoreType } from "@/interface";
 import { useCallback, useEffect } from "react";
