+++ src/components/StoreBox.tsx
@@ -1,4 +1,5 @@
-import { Dispatch, SetStateAction } from "react";
+"use client";
+
 import Image from "next/image";
 import {
     AiOutlineClose,
@@ -7,7 +8,7 @@ import {
     AiOutlinePhone,
 } from "react-icons/ai";
 import { HiOutlineMapPin } from "react-icons/hi2";
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
 import { useRecoilState } from "recoil";
 import { currentStoreState } from "@/atom";
 import Like from "./Like";
