+++ src/components/StoreList.tsx
@@ -1,6 +1,8 @@
+"use client";
+
 import { StoreType } from "@/interface";
 import Image from "next/image";
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
 
 interface StoreListProps {
     store?: StoreType;
