+++ src/components/comments/index.tsx
@@ -1,7 +1,9 @@
+"use client";
+
 import { useSession } from "next-auth/react";
 import CommentForm from "./CommentForm";
 import CommentList from "./CommentList";
-import { useRouter } from "next/router";
+import { useRouter } from "next/navigation";
 import axios from "axios";
 import { CommentApiResponse } from "@/interface";
 import { useQuery } from "react-query";
@@ -9,13 +11,16 @@ import Pagination from "../Pagination";
 
 interface CommentProps {
     storeId: number;
+    params?: {
+        page: string;
+    };
 }
 
-export default function Comments({ storeId }: CommentProps) {
+export default function Comments({ storeId, params }: CommentProps) {
     const { status } = useSession();
 
     const router = useRouter();
-    const { page = "1" }: any = router.query;
+    const page = params?.page || "1";
 
     const fetchComments = async () => {
         const { data } = await axios.get(
