--- src/pages/_app.tsx
@@ -1,32 +1,0 @@
-import Layout from "@/components/Layout";
-import "@/styles/globals.css";
-import type { AppProps } from "next/app";
-import { QueryClient, QueryClientProvider } from "react-query";
-import { ReactQueryDevtools } from "react-query/devtools";
-import { SessionProvider } from "next-auth/react";
-import { RecoilRoot } from "recoil";
-
-import { ToastContainer } from "react-toastify";
-
-const queryClient = new QueryClient();
-
-export default function App({ Component, pageProps }: AppProps) {
-    const { session } = pageProps;
-    return (
-        <RecoilRoot>
-            <QueryClientProvider client={queryClient}>
-                <SessionProvider session={session}>
-                    <Layout>
-                        <Component {...pageProps} />
-                        <ToastContainer
-                            autoClose={1000}
-                            pauseOnFocusLoss={false}
-                            pauseOnHover={false}
-                        />
-                    </Layout>
-                    <ReactQueryDevtools />
-                </SessionProvider>
-            </QueryClientProvider>
-        </RecoilRoot>
-    );
-}
